feat(CB2-14459): update values from parent component
