+++ src/app/features/tech-record/components/tech-record-summary/tech-record-summary.component.ts
@@ -175,6 +175,29 @@ export class TechRecordSummaryComponent implements OnInit, OnDestroy, AfterViewI
 			.subscribe((techRecord) => {
 				if (this.isEditing && techRecord) this.form.patchValue({ ...techRecord });
 			});
+
+		// TODO clean this up in the future
+		const formControl = this.form.get('techRecord_vehicleConfiguration');
+		formControl?.valueChanges.pipe(takeUntil(this.destroy$)).subscribe((value) => {
+			if (value === 'articulated') {
+				this.form.patchValue({
+					techRecord_bodyType_description: 'articulated',
+					techRecord_bodyType_code: 'a',
+				});
+			}
+			if (value) {
+				const functionCodes: Record<string, string> = {
+					rigid: 'R',
+					articulated: 'A',
+					'semi-trailer': 'A',
+				};
+
+				const functionCode = functionCodes[value];
+				this.form.patchValue({
+					techRecord_functionCode: functionCode,
+				});
+			}
+		});
 	}
 
 	get vehicleType() {
